refactor: code-review 指摘 10 (caching<->dataset 双方向依存を utils への葉モジュール化で解消)
- IMAGE_TRANSFORMS / load_image / get_crop_ltrb / trim_and_resize_if_required
  を library/utils.py へ移動 (dataset.py / caching.py からは再 export で互換維持、
  BucketManager.get_crop_ltrb は委譲 staticmethod として存続)
- caching.py の _ds() 遅延アクセサを撤去し utils 直 import に
- strategy_base.py の関数内 `from library import caching` を top-level に昇格
- utils.py の未使用関数 swap_weight_devices / weighs_to_device を削除
  (custom_offloading_utils.py に使用中の別実装があり、こちらは孤立した古いコピー)
